Updated iloot.py to support "pass the auth token"
Added the "--auth" option. This option takes an X-MobileMe-AuthToken (which can be gained from MiTM of an iOS device), and bypasses the need for a username and password to download iCloud backups.

As part of this commit I've also changed the username/password options to be optional, rather than positional arguments.